+++ src/main/java/com/moler/task/repository/PointRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface PointRepository {
     List<Point> getAll();
     Point save(Point point);
-    Point getPointById(Long id);
+    Point getPointById(Integer id);
 
 }
